Make URI helpers standalone
